sw/python/requirements.txt
absl-py==0.12.0
aenum==3.1.0
astor==0.8.1
astunparse==1.6.3
atari-py==0.2.9
bitstring==3.1.7
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cloudpickle==1.6.0
cycler==0.10.0
eigenpy==2.6.4
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gym==0.18.3
h5py==3.1.0
idna==2.10
importlib-metadata==4.5.0
joblib==1.0.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mini-cheetah-motor-driver-socketcan==0.1
moteus==0.3.23
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyelftools==0.27
pyglet==1.5.15
pyparsing==2.4.7
pyserial==3.5
python-can==3.3.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
six==1.15.0
stable-baselines==2.10.2
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
